# First pass in simplifying and improving qmm