Extracted the loading and dequantizing logic to a reusable `QuantizedBlockLoader` à la `steel`. Also removed the reading of scales/biases in smem and reduced the tile size that provides a pretty big speedup.

Before on Mistral 7B 4bit 64 group size:

```
1000 token prompt: 880 tps
QLoRA validation: 12.8s
QLoRA training: ~533 tps 
```

after

```
1000 token prompt: 1100 tps
QLoRA validation: 9.97s
QLoRA training: ~653 tps
```

all in all a pretty consistent 20%-25% speedup.