@angeloskath 2024-04-24
After @jagrit06 's improvement to avoid memory bank conflicts and remove the checks out of the loop we now get 

```
1000 token prompt: 1230 tps
QLoRA validation: 8.99s
QLoRA training: ~725 tps
```

for a total improvement of 30%-40%!